+++ src/main/java/org/noostak/auth/application/KakaoServiceImpl.java
@@ -8,7 +8,7 @@ import org.noostak.auth.common.exception.AuthErrorCode;
 import org.noostak.auth.common.exception.AuthException;
 import org.noostak.auth.common.exception.ExternalApiException;
 import org.noostak.auth.domain.vo.AuthId;
-import org.noostak.auth.dto.*;
+import org.noostak.auth.dto.kakao.*;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
