+++ src/main/java/org/noostak/auth/dto/kakao/KakaoLogoutResponse.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.dto;
+package org.noostak.auth.dto.kakao;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
@@ -8,20 +8,19 @@ import lombok.Getter;
 import org.noostak.auth.common.exception.KakaoApiErrorCode;
 import org.noostak.auth.common.exception.KakaoApiException;
 
+
 @Getter
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoTokenInfoResponse {
-    private String id;
-    private String expires_in;
-    private String app_id;
+public class KakaoLogoutResponse {
     private String error;
     private String errorDescription;
     private String errorCode;
 
     public void validate(){
-        if(error!=null){
+        if(error!= null){
             throw new KakaoApiException(KakaoApiErrorCode.KAKAO_API_ERROR, errorDescription);
         }
     }
 }
+
