+++ src/main/java/org/noostak/auth/dto/kakao/KakaoUnlinkResponse.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.dto;
+package org.noostak.auth.dto.kakao;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
